requirements.txt
streamlit==1.38.0
google-genai==1.0.0
PyMuPDF==1.24.10
python-docx==1.1.2
python-dotenv==1.0.1
keybert==0.8.5
textblob==0.18.0
sentence-transformers==3.1.1
